+++ app/src/main/java/com/fibelatti/photowidget/home/MyWidgetsScreen.kt
@@ -103,9 +103,9 @@ fun MyWidgetsScreen(
                                 aspectRatio = widget.aspectRatio,
                                 shapeId = widget.shapeId,
                                 cornerRadius = widget.cornerRadius,
-                                opacity = widget.opacity,
                                 modifier = Modifier.fillMaxSize(),
-                                blackAndWhite = widget.blackAndWhite,
+                                opacity = widget.opacity,
+                                saturation = widget.saturation,
                                 border = widget.border,
                                 isLoading = widget.isLoading,
                             )