+++ app/src/main/java/com/fibelatti/photowidget/ui/ShapedPhoto.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.core.graphics.toColorInt
 import com.fibelatti.photowidget.model.LocalPhoto
+import com.fibelatti.photowidget.model.PhotoWidget
 import com.fibelatti.photowidget.model.PhotoWidgetAspectRatio
 import com.fibelatti.photowidget.model.PhotoWidgetBorder
 import com.fibelatti.photowidget.platform.getDynamicAttributeColor
@@ -21,9 +22,9 @@ fun ShapedPhoto(
     aspectRatio: PhotoWidgetAspectRatio,
     shapeId: String,
     cornerRadius: Int,
-    opacity: Float,
     modifier: Modifier = Modifier,
-    blackAndWhite: Boolean = false,
+    opacity: Float = PhotoWidget.DEFAULT_OPACITY,
+    saturation: Float = PhotoWidget.DEFAULT_SATURATION,
     border: PhotoWidgetBorder = PhotoWidgetBorder.None,
     badge: @Composable BoxScope.() -> Unit = {},
     isLoading: Boolean = false,
@@ -41,7 +42,7 @@ fun ShapedPhoto(
             shapeId,
             cornerRadius,
             opacity,
-            blackAndWhite,
+            saturation,
             border,
         ),
         isLoading = isLoading,
@@ -61,7 +62,7 @@ fun ShapedPhoto(
                 bitmap.withPolygonalShape(
                     shapeId = shapeId,
                     opacity = opacity,
-                    blackAndWhite = blackAndWhite,
+                    saturation = saturation,
                     borderColor = borderColor,
                     borderPercent = borderPercent,
                 )
@@ -70,7 +71,7 @@ fun ShapedPhoto(
                     aspectRatio = aspectRatio,
                     radius = cornerRadius * localDensity,
                     opacity = opacity,
-                    blackAndWhite = blackAndWhite,
+                    saturation = saturation,
                     borderColor = borderColor,
                     borderPercent = borderPercent,
                 )