+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+* Replaced the black and white toggle with a saturation picker for granular
+  control
+
 ### Fixed
 
 * Fix the behavior of tapping to view the previous photo when shuffle is on